completed the diagram except for the write transaction part
I converted c code into a hlsm. when I am drawing the part about the write transaction, I realized that some state sequences (the ones I added for read and write tr.) are repeating. I thought that I can functionalize them. I did it for the write transaction (the read operations was designed before I came up with this idea) as a template. didn't fill the 'function body' of the write transaction yet.